LT-19843: Remove duplicate refresh of Reversal Indexes
Removed the “MasterRefresh” that was being done in
ReversalListener.ChangeOwningObject(). This was resulting in a call
to XhtmlDocView.UpdateContent(), which is already being called
when we get the FwXWindow.OnPropertyChanged() notification for
currentContentControl. These both were starting background threads
to interact with the PublishedHtml file.
The MasterRefresh call was added with commit 1e3a4c4e, to fix
LT-18790.
This submission adds a different way to fix LT-18790; we no longer
"SuspendLoadingRecordUntilOnJumpToRecord" for HistoryBack and
HistoryForward.

These changes also improve performance by about 40% when switching
to the Reversal index since we no longer build the html file twice
and update the display twice.

Change-Id: I60fce6f081de5465c0a6a2f93df3ff360fc4f450