LT-20952: Reload the Reversal Index after adding
Adding a single Reversal Index Entry in certain cases (at the end of
the list for a sense, in certain projects) hit a special case in
ReloadList that seems to be peculiar to updating or replacing a single
item. To resolve this, reload the whole list when a new item is added.

Refactor RecordList.ReloadList(int, int, int) to improve
performance and readability. Unit test.

Change-Id: I627bee15e84ba94ddc563073b68afcb099ce8b45