LT-18315: Fix problem with using active clerk
When uploading to Webonary we were passing the active clerk to the
method used to generate the reversal letter headers.  We now pass
in the reversal clerk for all the reversals.

Notes:
- Inside ExportConfiguredReversalJson() the active clerk is
temporarily switched to the reversal clerk and then restored to
the original clerk. The call to GenerateReversalLetterHeaders()
is a few lines after the call to ExportConfiguredReversalJson().

- This problem was found using the steps listed in LT-21097.

Change-Id: I820e6bff32ad7e9e595ddc604c3d782c78c63896