LT-21198: Webonary upload – Use reversal writing systems
When generating data for reversals that are not currently
displayed we need to use the reversal WritingSystem instead
of the writing system for the displayed column.
Set an override to use the reversal WritingSystem.

Also, removed some asserts that were firing for a valid case
when getting the text for the page buttons. (I was the person
that added these asserts a couple months ago.)

Change-Id: I540ba8d68ab66c6fc8c3fc86d5ff58bbe1bc24a4