Remove old FdoUi files
These files have either completely or partially moved and the
new files seem to have all the changes that were made in 9.1.
No additional edits to the new files appear to be needed.

  FdoUi\FwLcmUI.cs
  moved to
  FwUtils\FwLcmUI.cs

  FdoUi\LexEntryUi.cs
  moved to
  LanguageExplorer\Controls\LexEntryVc.cs

  FdoUi\Dialogs\SummaryDialogForm.cs
  moved to
  LanguageExplorer\Controls\SummaryDialogForm.cs

This test file does not exist in the new file structure
and the SUT no longer exists. So, it should be ok to delete.
  FdoUi\FdoUiTests\FdoUiTests.cs

Change-Id: I74a8879e342fa3656158eaac43194ca03a96b5cc