Remove old XCore files
This file has edits that were made in 9.1 with Commit
18595c33 that are needed.
  XCore\xCoreInterfaces\PropertyTable.cs
  moved to
  LanguageExplorer\Impls\PropertyTable.cs

These files have either completely or partially moved and the
new files seem to have all the changes that were made in 9.1.
No additional edits to the new files appear to be needed.

  XCore\xCoreInterfaces\xCoreInterfacesTests\PropertyTableTests.cs
  moved to
  LanguageExplorerTests\Impls\PropertyTableTests.cs

  XCore\xCoreTests\InventoryTests.cs
  moved to
  LanguageExplorerTests\InventoryTests.cs

This test file does not exist in the new file structure
and the SUT no longer exists. So, it should be ok to delete.
  XCore\xCoreTests\IncludeXmlTests.cs

The class in this file does not exist in the new file
structure. So, ok to delete.
  XCore\xCoreInterfaces\Mediator.cs

Change-Id: I56369f505ca4157716f10367144d2dd5d525c502