Fix Exception when restoring backups
Restoring some backup files (LT-21198) results in the exception:
“FieldWorks ran into a problem trying to display this object in
DataTree::ApplyLayout: Requested value 'mnuDataTree-VariantForms'
was not found.”

The problem is that 'mnuDataTree-VariantForms' is stored in the
backup file, but was changed to use an underscore instead of
using a dash: ‘mnuDataTree_VariantForms’.

To avoid doing a data migration this PR changes back to using
names with dashes.


Remaining TODO:
I need to do a more thorough search through the existing files to
see if there are other hardcoded strings with underscores that
need to also be changed back to strings with dashes.

Change-Id: If9569e1caf01962db5ef3f27b0f0bebe6016bdaa