LT-21294: handle stem names correctly (XAmple)
Word grammar does not handle stem names correctly for a template that
requires more derivation. The XAmple word grammar uses a (PC-PATR)
logical constraint to check for the validity of stem names used in an
affix template.  For normal affix templates, the logical constraint
uses 'Full' and 'Stem' for the appropriate non-terminal pieces of the
constraint.  For affix templates that require more derivation, the
non-terminal pieces should be 'Stem_1' and 'Stem_2', respectively.
They were still 'Full' and 'Stem' so the logical constraint did not
apply correctly.

Adjust indentation for one line in the transform and the expected test
output

Change-Id: Iac36feb5e54c4a8fbef51240e1e34f3b0035227b