LT-21265: Fix the second exception when viewing Bulk Edit
This change fixes the second of three exceptions when viewing
Bulk Edit. The exception is listed below.

The code was using the flid of the elements on the list and
assigning that value to the flid of the list, which resulted in
the “does not match” exception. To fix this problem:
1. Restored from 9.1 the GetFlidOfVectorFromName() method in
   RecordList. The new version is mostly the same, with some
   functionality removed that was not needed for this fix.
2. Wrote a new GetSourceFieldName() method that provides the
   same information as the 9.1
   PartOwnershipTree.GetSourceFieldName() method, but uses
   direct lookups instead of getting the information from xml.

Note:
It looks like the problem was introduced in the following
commit where the ReloadList() method moved from RecordList.cs
line 693 to EntitiesOrChildClassesRecordList.cs line 221:
  commit 6c280cd1
  09/16/2015 05:02 PM
Prior to this commit the code was almost the same as 9.1. During
this move the following line of code was added:
	m_flid = newTargetFlid;

Exception:
Msg: for some reason BulkEditBar.ExpectedListItemsClassId(5002)
 does not match SortItemProvider.ListItemsClass(5035).
Class: System.ApplicationException
Source: LanguageExplorer
Assembly: LanguageExplorer, Version=10.0.0.0, Culture=neutral,
 PublicKeyToken=null
Stack:    at LanguageExplorer.RecordList.
  CheckExpectedListItemsClassInSync(Int32 beExpectedListItemsClass,
  Int32 recordListExpectedListItemsClass) in C:\fwrepo\fw\Src\
  LanguageExplorer\RecordList.cs:line 1554
at LanguageExplorer.Areas.Lexicon.Tools.
  BulkEditEntriesOrSensesTool.EntriesOrChildClassesRecordList.
  ReloadList(Int32 newListItemsClass, Int32 newTargetFlid,
  Boolean force) in C:\fwrepo\fw\Src\LanguageExplorer\Areas\
  Lexicon\Tools\BulkEditEntriesOrSensesTool.cs:line 463
at LanguageExplorer.RecordList.OnChangeListItemsClass(Int32
  listItemsClass, Int32 newTargetFlid, Boolean force) in
  C:\fwrepo\fw\Src\LanguageExplorer\RecordList.cs:line 1010
at LanguageExplorer.Areas.RecordBrowseView.TargetColumnChanged(
  Object sender, TargetColumnChangedEventArgs e) in C:\fwrepo\fw\
  Src\LanguageExplorer\Areas\RecordBrowseView.cs:line 257

Change-Id: Ib9ec7056a2af031428ecfd221c928ac5e645f134