LT-21295: Upload ProjectDictionaryOverrides.css to Webonary
Three problems were preventing the ProjectDictionaryOverrides.css
file from being uploaded to Webonary:
1. The project path passed in to CopyCustomCssAndGetPath() contained
the file name of the .fwdictconfig file. It should have been a path
without the file name.
2. We were not removing the path from the
ProjectDictionaryOverrides.css file name when adding it to
dictionaryMetaData.mainLanguage.cssFiles. This resulted in us
posting to Webonary the full path along with the file name, in a
field that should just contain the file name.
3. If the user never configured the dictionary for the project, then
the project path passed in to CopyCustomCssAndGetPath() was
incorrect.  Instead of the path to the project folder, we were
passing it the default configuration folder
(…\Language Explorer\DefaultConfigurations\Dictionary). This was
the default location of the .fwdictconfig file.

Note: There were no problems with uploading
ProjectReversalOverrides.css.

Change-Id: Iec3f722f75054e49607483b035d4ce45855d0f4c