LT-21360: Fix crash when switching templates
This change moves the code that calls ConstChartBody.SetRoot()
so that it is called before rebuilding the headers. In some
cases rebuilding the headers results in calls to the body, so
the body data needs to be up to date. In this case the old column
information was causing the crash.

Change-Id: I934ba31b9cb871c4303e6171dd953af75998391b