LT-21310: Fix problem with custom fields that have a suffix
The way we were keeping track of which custom fields need to be
added to the parent was causing us to remove custom fields when
there was more than one instance. (All of the copied instances
have a suffix.) To prevent this we no longer modify the original
list of custom fields. Now there is an additional list that we
can modify.
The test that I modified was previously passing because
the old call to customFieldNodes.Remove(configNode) would
not remove custom nodes with a suffix. If the order of
the custom nodes were switched, then the test would have
failed.

Change-Id: Ia9d4932db004cc0375969538b69f41d08ad0b395