LT-21308: Display enabled writing systems
Do not hide writing systems that are enabled.  Also changed
a tooltip.

Notes for future development:
In Feb 2017 this code was originally added with Commit 577e4293
and it looks like it would have displayed the enabled ws’s.
It appears to have regressed with Commit 5d822c97, a month and a
half later.

There are currently two situations where we do not preserve
the user specified order:
1. The order of the ws’s that are not enabled is currently
not preserved.
2. Instead of preserving the user order we group the
enabled ws’s to all be immediately after the default(s).
The code in UpdateWsOptions() and LoadAvailableWsList() could
be cleaned up if we stopped supporting these behaviors (it could
also be cleaned up to a lesser extent if we continued to support
them).

Change-Id: I87466c0633fdd28125dca1d4ea4e79238be299f3