LT-21433: Fix crash when ICU collation starts with a comment
There is code that treats a collection that starts with a ‘#’
(a comment) as a special case. Possibly that code existed before
comments were allowed on collections. If what followed the ‘#’ was
not in the format expected by the special case, then we threw an
exception. Now, instead of throwing, we let the string fall through
to the normal processing.

Change-Id: I158b83900d228e963489c2408c072f6192cb22ce