LT-21435: Performance improvement when parsing with XAmple
I am submitting changes made by Andy Black. Here are comments from
Andy:
Realizing that this is a suffixing-only language led to refining
the XAmple input files to catch bad category mappings early instead
of waiting for the Word Grammar to find them.  We had designed the
XAmple parser with languages which have both prefixes and suffixes
in mind.  If the language has no prefixes (or infixes), then there
are things we can do to speed up the parsing.

Change-Id: Ic2369a6641243ac0965746f72290e1710b3720b4