LT-21568: Cleanup WritingSystemHvo property change
Did not add the cleanup to RawTextPane.cs because there are other
code paths in EditingHelper::WritingSystemHvoChanged() that could
results in returning without changing the writing system.

Note:
In addition to cleanup this change also fixed a bug that was
introduced in Commit eeb9f099. The defect was because
m_fSuppressNextWritingSystemHvoChanged was being set to true after
the SetProperty() call in EditingHelper.cs.  This was ok when we were
processing the property change OnIdle, but did not work when commit
eeb9f099 change the property change handling to be immediate.

Change-Id: I9be87f0af9955abac657f0b69227d9706665818b