LT-21492: Remove WritingSystemUpdated & WritingSystemDeleted
Removed the unused code (LT-21492 and LT-21493).
Only test code was passing a mediator in to the
FwWritingSystemSetupModel constructor, so only test code
was calling SendMessage for both of these events.
The two delete tests that were failing an assert when I removed
the mediator both contain other asserts that still fail if the
writing system is not deleted.

Change-Id: I3aab4e1ce81e6ad888ceed19c15c0c778691017a