LT-21607: Fix dialog flash selecting diff Entry or Reversal
Remove blank page case from LoadPageIfNecessary.

Page button is null for projects small enough that all entries fit
on one page, and the blank page case of LoadPageIfNecessary causes
a full refresh for such projects anytime a new entry is selected.

This case is not needed, since blank page occurs when FLEx is
just being opened or is switching between areas.
- When FLEx is just being opened, document is null, so
LoadPageIfNecessary will not execute. Instead, XhtmlDocView's
PostLayoutInit handles loading the Dict or RevIndexes view
for the 1st time.

- When FLEx switches between areas, ActiveClerkSelectedObject
is not changed so LoadPageIfNecessary will not be called,
and loading in this case is also handled elsewhere.
Change-Id: If0120d9184832aebafc5307adf7823cebc51110f