LT-21608: Execute ‘End of Action’ events in a fixed order
This change fixes LT-21608 by moving the RecordNavigation event
so that it executes at the end of the action, instead of
immediately. It also establishing a fixed order for executing the
RecordNavigation event before the SelectionChanged
event, regardless of which one gets queued up first.

This pattern should be used for other events that need to
execute at the end of an action.

Follow-up tasks:
1. The property changes that result in calls to
   PropTable.HandlePropertyChangeOnIdle() need to
   be moved to the EndOfActionManager.
2. Review the remaining mediator property changes, mediator
   events, and IdleQueue events to try and determine which may
   need to execute at the end of the action.
3. Fix a remaining display problem after executing the steps in
   LT-21608 where the Texts area does not display the
   characters that are typed in the Title area.

It looks like the following defects might also be fixed with
this change:
LT-21606
LT-21599
LT-21600
LT-21613
LT-21602
LT-21605

Change-Id: I9ee5a4d6eb3ef749088922dcf3b14e7285f4fc01