LT-21630: EndOfAction events no longer re-add as EndOfAction
The executing code now checks to see if an EndOfAction event
is currently executing to determine if the event should execute
immediately or at the EndOfAction.

An exception is thrown if a EndOfAction event is added and it
does not come after the currently executing event.


Notes:
The changes to move CreateFirstRecord to it’s own event are probably
not needed to fix LT-21630.  They are the result of a different
attempt to solve LT-21630 that didn’t fully solve the problem.
These changes are still helpful in simplifying what happens in
EndOfAction execution.

In XmlBrowseViewBase.cs SelectedIndex.Set the two Publish calls were
changed to not be EndOfAction.  These changes are not needed to fix
LT-21630. They were to reduce our use of EndOfAction and could
be changed back to EndOfAction if needed.

The check for IsDisposed in XWorksViewBase.cs ExistingClerk is to
fix another exception that is thrown after these changes fix the
exception that is thrown in LT-21621.  The changes in Commit eb6b2908
moved the call to UpdateClerkEditable.

The exceptions reported in these defects are also fixed with
this change:
LT-21621 	No exception, but there is an assert if the last 2 steps
          are repeated a second time.
LT-21624	No exception, but the navigation to ‘th’ does not work.
LT-21627	Fixed

Change-Id: I06d1ebffc975d6d1ab2c7b2b50f9b4d9947b2cc4