LT-21643: Fixed Failed Jump for Back Button
The following exception was being thrown from
LinkListener.FollowActiveLink() at the line after the
m_mediator.SendMessage("SetToolFromName",...) call.
    System.NullReferenceException: 'Object reference not set to
    an instance of an object.'
    m_lnkActive was null.
The SendMessage call resulted in a SetProperty call for
"areaChoice". This property has been moved to PubSub so the
subscribers execute immediately (instead of OnIdle), which
results in m_lnkActive being set to null in LinkListener.

Change-Id: I577a6b7d2da8fa9b800566f145930292320c5ea6