Added cmd capability to clone and open a project
When launched from CMD with the proper arguments, flex will
now interface with flexbridge to obtain a specific project
with minimal user interaction. Currently, flexbridge
confirms with the user before downloading
from the given URI using a simple prompt.

This change required a number of new parameters
to be passed to flexbridge, and we opted to overload
the launching method,
to help preserve functionality for any external callers.

I also changed the Process.Start logic
slightly in order to deliver the user/pass
to the process over environment variable,
as opposed to less secure solutions.

This work is mostly done, but it doesn't support
custom project names yet.
And I still need to consider input checking for the new CMD arguments.

Ticket link for FB: https://github.com/sillsdev/flexbridge/pull/388

Change-Id: Ib4a85a628216f382780aaa50f2e70bd720ef00a2