LT-21648: Add “Caption (or Headword)” to Dictionary Config
In Dictionary Configuration->Pictures add an option that will first
try to get the Caption for a writing system. If there isn’t a
caption then try to get the Headword for the writing system.

Note:
In the fwdictconfig files we also made changes to Pictures->Headword
so that they display the Headwords for the specified writing system.
Previously we were always displaying the headword for the default
ws, regardless of which ws the user selected. If multipe ws’s
were selected we were displaying the default headword multiple
times.

Change-Id: I91d66eebaba69a832a30c2fafe6c452e9b6dbf27