LT-21443: ‘Approve All’ - mark as user-approved
Marking the analysis as user-approved is a partial fix for this
Defect.
The following problems still exist when testing with the project
‘Parser-Test2 Approve all 2024-01-16’:

Step 5: There are 2 analysis, both in User Approved.  The first
has *** for Lex Gloss, the second has ‘I + NOM’ for Lex Gloss.
(Note: This may be an issue because the capital word Sayana is
also in the text.)

Step 7: Both Analysis are shown as ‘User Opinion Unknown’

Step 9: Both Analysis are displayed in the correct section, but
the one in ‘User Approved’ displays ‘***’ instead of displaying
‘he’ for the Lex Gloss.
(Note: If in the ‘Word Analyses’ view you left click on the
Lex Gloss that is displaying the ***, then it populates with
‘he + NOM’ and if you go back to the text, then ‘he + NOM’ is
displayed in the Lex Gloss field for both ‘diana’ words, even
if it’s not selected.)

Step 10: When not selected Lex Gloss is displayed as ‘***’,
Except for the first word in row 1. On row 1 the capital word
‘Sayana’ shows the Lex Gloss value even when not selected
(I + NOM).

NOTE: If I Use ‘Approve Throughout this Text’ for all the
words, then none of these problems exist.  Everything displays
as described in the defects Acceptance Tests.


Change-Id: Ic19d9391aa364016c63e2ab766ba4b2eab69965f