LT-21424: Fix Approve undo issue
Fixed the undo problem described in Scenario 4 in the defect.
The undo problem only existed after selecting a Morphemes and
then clicking on another word to do the approval. (The other
approval workflows did not have this issue.) The problem
was resolved by executing the code in a way that is similar to
what the other approval workflows do.

Change-Id: Ia62e5a8be4c503ce4a4b972b9ec32e335b96b24b