LT-21672: Add Tables to Word Export
This adds the basic table data, without styling. Styling
will be added in a separate change.
Tables cannot be added to a paragraph, they need to be added
directly to a Body. So, the majority of this change was to
not create paragraphs until the end of the process; when we
are in AddEntryData().
What previously was a single paragraph for
the Entry, could now be multiple paragraphs if there are one or
more tables not at the end of the Entry.

Change-Id: I4923a87c014c2c4f51159d37d22404fceb64e5f9