LT-21672: Add styles to table cells
This adds support for when a style is applied to a specific
cell or when different styles are applied to different pieces
of the text in a cell.

Note: The changes in SetRunStyle() also add support for styles
that are applied to all or a portion of the text in other
fields (notes, custom fields …).

Change-Id: Id1ca7ccfaeaad7e7ea7e781263151b1714d9d90f