LT-21761: Fix indenting for the first line after a Table (#37)
* LT-21761: Fix indenting for the first line after a Table

The first line of an entry typically uses FirstLineIndent to
indented to the left. When we have multiple paragraphs for an
entry (because the entry contains Tables or Pictures)
then we do not want the line following the Table or Picture to
also be indented to the left. We now create a ‘continuation’
style for paragraphs following a table or picture.  It is the
same as the original style except it does not use FirstLineIndent.