Disable borders in Word Export
- Comment out handling of borders in Word Export

Borders do not currently display in FLEx.
But once a border is added to a style in FLEx, deselecting
the border does not actually remove the border object from
the FLEx style. This means if a border has ever been added
in FLEx, it will still exist in the configuration node
and be added in the word export. Until this is fixed, we
should disable borders in the word export.

Change-Id: I19063eaaa9bb85f110ddb4c177958a7887e3af8c