LT-21778: Fix image textframe styles
Fixed the problem with the continuation style being written to
paragraphs containing images/captions (instead of the textframe
style being written).
To prevent the problem with the paragraph following the
image/caption from using the textframe style, we now base the
continuation style on the first paragraph instead of the last.

Change-Id: Ibae531ebf142719d55b912efe72746299a1ce642