LT-21771: Word Export – support indenting for groups
This change will also fix indenting for any other node type
that has an ancestor with indenting. Which doesn’t seem common,
usually it’s zero or the ancestor does not have a paragraph style.
Notes:
- The only line of code in WordStylesGenerator.cs that changes
logic is in CalculateMarginLeft() where I changed:
leadingIndent -= ancestorMargin + hangingIndent;
to:
leadingIndent -= hangingIndent;

The rest of the changes are all cleanup, most of then related to
no longer needing to calculate ancestorMargin.

- The AddStyles() call that was removed from GenerateGroupingNode()
was a duplicate call.

Change-Id: I1a1301c5b5444056119c915816f0c2a66d861a6a