LT-21785: Add more style support (#77)
* LT-21785: Add more style support

- When a run is created set the style name. (It may get modified
  with basedOn information as we get other notifications.)
- Add styles during run creation.
- Allow styles to be based on other styles, which can be based on
  other styles…
- Support different styles on different document fragments.

TODO:
A remaining task is to change the style names from the
config.Style to the config.DisplayLabel.  This will improve
the style names that appear in Word. At that time we will need
to add additional code to generate a unique style name if the
properties of a style are different.