Fix 2-column display by correcting image anchor
The horizontal anchor type for the textframe should be set to text
rather than page.

A text type anchor promts the textframe to move with the text it is
anchored to, while a page type anchor would keep the textframe
fixed in a particular location on the page.

If the horizontal anchor type is set to page, images can appear
in the wrong column when the document layout is set to two columns.

Change-Id: I28c1644d0ce7bdf33d80d908d3615f12b8a059ab