Word Export LT-21942: Fix picture clipping problem (#177)
Runs containing drawings (pictures) need to be in their own
Paragraph. When we are in SeparateIntoFirstLevelElements()
don’t combine them with other elements in a paragraph.  Instead
leave the run as a first level element so that later, in
AddEntryData(), it will get put in its own paragraph with the
correct style.

Change-Id: Ie62eb2edb7547d553018a61ea57b638dc146d407