Word Export: prevent cross-thread exception
When disposing the ClerkActivator do not update the status bar
in the call to ActivateUI().
XHTML export is also calling this from a different thread but we
were not seeing this error because the text string was the same
as the current value so no change was made.
Notes:
In OnExport(), after the command completes, we call
ActivateUI() a second time, and this time we do update the
status bar.
The exception was only happening when there were reversals
and it seemed to only/mostly happen in debug builds.

Change-Id: Ife8957181e5a07736b1f56e2f488aa92c8f226c5