Word Export: writing system that is not part of the project
Prevent Word Export from crashing when we encounter a string
that uses a writing system that is not part of the current
FLEX project. This can happen when using copy/paste from one
FLEX project to another, and the projects use different
writing systems. See LT-21981.
Pasting table data from a custom field in one project to a
custom field in another project reproduced the crash.

Change-Id: Ic238b27a83e9591bc81eac9bdf4628ab11787352