LT-21882: Webonary: Fix style names
References to styles in the xhtml were not matching up with the
correct style names in the css.

- In the Json generator we were not using the _styleDictionary to
generate the css.  Instead we were re-generating the css from the
configuration. Now use the _styleDictionary to generate the css.
- In the Json generator there were two _styleDictionaries being
used. They were filled at different times and could have
different values because of unique name generation. Now use
one _styleDictionary.
- In many situations we were still generating styles with unique
names when an existing style was identical.  Changed the code to
re-use existing styles that had unique names. (This change was
not necessary to fix the problems described in this defect but
resulted in a substantial size reduction in the css file size.
The Jii-entries project that is attached to the defect, and the
Sena 3 project both result in a css file that is less than 10%
of the old size.)

Change-Id: I457674ca44b37ed966cb1d8995011a0a19df4cca