Word Export: LT-22006: RTL - Fix indentation and clipping
Word does not expect the indentation direction (left/right) to
be flipped for right to left output.  Removed the extra code
and output the same indentation direction for both LTR and RTL.

This problem also existed with the single column output, but
became more obvious when we switched to double column output.

Change-Id: Id86359ec0524f0eca8ef3c6e4aaa176efcac5122