Word Export: LT-22051: Fix problems with incorrect styles
Change the way that we generate character styles so that it
includes styles from all the parent nodes. Also changed the
way paragraph styles are generates so they no longer directly
include run properties.  Now they are linked to a character
style that will be used to build the character style properties.

Change-Id: I04921ad04efed617a31020616357195f9a12cd1e