LT-22099: Fix XHTML and Word Style problems
We were not displaying the correct styles for Subentries and
Subsenses (nodes that are shared). The problem was that we were
using the style from the parent node, when we needed to use the
style from the node that was making the reference.  This change
keeps a node list, which allows us to collect all the style
information from where the node is used, not necessarily from
the parent node.

Change-Id: Id8c2adbc7c641831be802e61bfec834d61b616c9