LT-22099: Fix XHTML and Word Style problems (#326)
We were not displaying the correct styles for Subentries and
Subsenses (nodes that are shared). The problem was that we were
using the style from the parent node, when we needed to use the
style from the node that was making the reference.  This change
keeps a node list, which allows us to collect all the style
information from where the node is used, not necessarily from
the parent node.

Note:
The nodeList contains the correct node information that we need to support subentries and subsenses, so we no longer need special code to fix these situations.  Subentries and subsenses use shared nodes (ie. nodes that are references from somewhere other than node.Parent). Because of the way nodeList is built, it will contain the correct node information for a shared node.  The key difference between nodeList and a list of nodes we could generate by using node.Parent is when a node is referenced (used) by a node other than it's parent. Here is an example:
If node.Parent resulted in this set of nodes:
A->B->C
And node C is a shared node that is also referenced from node E:
D->E->C
Then because of the way nodeList is built, it will contain D->E->C, so we no longer need to do special case code to try and switch from A->B->C to D->E->C.