Word Export: LT-22106: Leave ws tag on styles
If a style with a ws tag was identical to the same style
without the ws tag (the default) then in an effort to reduce
the number of styles in the Word document we were changing
all references to the style with the ws tag to the style name
without the ws tag. That optimization was undesirable so we now
leave the style with the ws tag.

Change-Id: I7dbdc54c2581bc5cbcc46eaab7be65b7901836f5