LT-21766: Fix Reversal Sort Order in Webonary
In some situations the CurrentReversalWsId is not correct.
Now we first try to get a ws from the SubComparer (if it is
a WritingSystemComparer). This looks like the same place that
the Sorter gets the ws.

Note: This change also fixes a similar problem with the reversal
sort order in Word Export.

Change-Id: I57cb30f7ce7f8cbe64958c659bc4c73609e53729