LT-21766: Fix Reversal Sort Order in Webonary (#363)
In some situations the CurrentReversalWsId is not correct.
Now we first try to get a ws from the SubComparer (if it is
a WritingSystemComparer). This looks like the same place that
the Sorter gets the ws.

Note: This change also fixes a similar problem with the reversal
sort order in Word Export.