LT-22122: Word Export performance improvement #2
There was data in a DocFragment that was only needed by the
fragment that was going to write to the docx file. We now moved
that data to a MasterDocFragment and only have a Body in the
regular DocFragment. Regular DocFragments are frequently created
and discarded.  Not having them open a memory stream saves a
significant amount of time.  Combined with the performance
improvements in commit e1943fb6, export times are now 30% to
50% of previous times.

Change-Id: I9ec6fa92709d3562c9331ec30c4c2920a71468b4