LT-21960: Fix export issues when there is filtering (#417)
When there is filtering applied to the entries, then only the
currently viewed entries would export correctly.  (If the main
dictionary was being viewed, then the filtered reversal entries
would not export correctly. If a reversal was currently being
viewed, then a filtered main dictionary would not export
correctly.)
The same problem existed for both Word and Xhtml exports. To
resolve the problem with filtered entries we now do the
collection of entries on the main thread.