LT-22124: Fix Sense homographs in Lexical Relations (#437)
When getting a property value for a field on a ISenseOrEntry
object, call the new method added to LCM that determines
the specific object (Sense or Entry), and the specific field
on that object, that should be queried to get the value.

In this specific defect we are trying to get the HeadWordRef
and the ISenseOrEntry is a LexSense. The specific field remains
HeadWordRef, but the specific object we need to get the value
from is the associated LexEntry. This is because the LexEntry
adjusts the HeadWordRef based on the decorator, the LexSense
does not.