LT-21960: Fix export for multiple reversals and filtering (#464)
When exporting to Word, Webonary, or XHTML, we now get the
entry data from direct data access, instead of relying on the
GUI to display the data we want to get.
The GUI updates were mostly being done through property change
notifications or mediator messages.  The problems we were seeing
was when these notifications and messages were combined with the
many existing special cases where we don’t want to update the GUI,
or we want to delay the GUI update. Getting the combination
correct to fix one situation was resulting in other situations
that were incorrect.

Problems resolved:
- Exporting multiple reversals to Word or Webonary.
- Exporting multiple reversals with different filtering
to Word or Webonary.
- Exporting to Word, Webonary, or XHTML immediately after starting
Flex. In this state, depending on what is being viewed, either the
main dictionary clerk or the reversal clerk has been created, but
not both.
- The GUI no longer flash with visual updates while doing an export.

Notes:
Since now we are getting the entry text without calling
XmlBrowseViewBaseVc.DisplayCell(), we no longer need the changes
that were added to fix LT-21198. This PR also removes the change
in Commit 0503457b that added OverrideWs.