Include all frameworks in local LCM, Palaso & Chorus
One of the LCM target frameworks is netstandard2.0. To build a
local LCM that uses a local Palaso, LCM needs the netstandard2.0
framework to be in the palaso nupkg.  By removing the
TargetFrameworks parameter from the ‘dotnet pack’ command the
nuget packages will contain all the frameworks specified in
the .csproj files.

Change-Id: Ia31076ec00105f880b88c975a8eee37dfa63d8b9