LT-3646: Unchecked checkboxes count as "no data"
When determining field visibility, treat unchecked checkboxes
as if they have "no data".

Note:  This change applies to all checkboxes stored as a Boolean,
and one checkbox that is stored as an integer.  To apply the new
behavior to a checkbox stored as an integer, and avoid changing
the behavior of all integer fields, we specifically identify the
integer field. There is a possibility that we missed other checkbox
fields that are stored as an integer.

Change-Id: Ia039b97f42727314b54d8d0b0303b7ca141b33e4