LT-22096: Keep focus field in view after reorder
There were 2 problems preventing the focus field from remaining
in view:

1.	We were doing a reference compare for the XmlElements:
slice.CallerNode == xnCaller
In the move-up and move-down workflows we are always
creating new slices; and the CallerNode is one of the created
slices. So the reference compare was failing and we were
never setting CurrentSlice.
I changed this so that we now compare the OuterXml values.
Note: Even if differentObject is set to false, we still
create new slices, at least in this particular case.

2.	Scrolling the control into view needs to be done after we
resume the layout.

Change-Id: Ie4ff0d0846ce50a65a57a9df515bbd06a16945d0