LT-22309: Fix RAMP error loading System.IO.Compression
The latest System.IO.Compression nuget package (version 4.3.0)
contains version 4.1.2.0 of System.IO.Compression.dll.  Palaso
and Flex are both using version 4.2.0.0 of the dll. From what I’ve
read it sounds like 4.2.0.0 is provided by the .NET Framework 4.6.2
or the runtime.
Adding the bindingRedirect resolved the error.

Note: I was not seeing the problem on a debug build, it was only
happening on a release build.